Lots of background stuff done
All functions now start with a capital letter, this is so it is easier
to differentiate functions and variables. There was a problem with all
classes that inherited the GameObject class, because a lot of the
variables were not initialized. I am going to search through all files
and init any possible variables. (obj_saw_bar depth is currently wrong,
it gets drawn behind the player)