Save Load Major code redesign
Saving and loading has been added. New class named GameObjectmanager. It
replaces the main_funcs_vars files. All object pointers are now private.
Including the file is not the most beautiful way, but it works... It
needs to be included in the .cpp file if it is inherited of GameObject.