Even more optimizing and lots of code redesign
A lot of the #includes in classes are moved from the .h file to the .cpp
file, this is because the files that include a class, don't necessarily
need all the files that class needs. The blood particles have been more
optimized. Particles that go outside of the room get destroyed.